sample_space and brainmapper readers read only registration output and scale it without loading the original data. User should load original data separately
